requirements.txt
attrs==21.4.0; python_version >= "3.7" and python_full_version < "3.0.0" and python_version < "4.0" or python_version >= "3.7" and python_version < "4.0" and python_full_version >= "3.5.0"
aws-cdk-lib==2.30.0; python_version >= "3.7" and python_version < "4.0"
aws-solutions-constructs.aws-lambda-stepfunctions==2.10.0; python_version >= "3.7" and python_version < "4.0"
aws-solutions-constructs.core==2.10.0; python_version >= "3.7" and python_version < "4.0"
cattrs==22.1.0; python_version >= "3.7" and python_version < "4.0"
constructs==10.1.43; python_version >= "3.7" and python_version < "4.0"
exceptiongroup==1.0.0rc8; python_version >= "3.7"
jsii==1.61.0; python_version >= "3.7" and python_version < "4.0"
publication==0.0.3; python_version >= "3.7" and python_version < "4.0"
python-dateutil==2.8.2; python_version >= "3.7" and python_full_version < "3.0.0" and python_version < "4.0" or python_version >= "3.7" and python_version < "4.0" and python_full_version >= "3.3.0"
six==1.16.0; python_version >= "3.7" and python_full_version < "3.0.0" and python_version < "4.0" or python_version >= "3.7" and python_version < "4.0" and python_full_version >= "3.3.0"
typing-extensions==4.3.0; python_version >= "3.7" and python_version < "4.0"
